Mega refactor: Change project structure
- Move models and utils into one single module: core
- Separate interfaces module into persistence interfaces and services interfaces modules
    - This allows to make a better separation of layers, as upper layers must only depend on the lower layer interfaces module
