+++ pom.xml
@@ -20,9 +20,9 @@
         <!-- ********************** -->
         <!--    Sub modules names   -->
         <!-- ********************** -->
-        <utils-module>spring-boot-example-utils</utils-module>
-        <models-module>spring-boot-example-models</models-module>
-        <interfaces-module>spring-boot-example-interfaces</interfaces-module>
+        <core-module>spring-boot-example-core</core-module>
+        <persistence-interfaces-module>spring-boot-example-persistence-interfaces</persistence-interfaces-module>
+        <services-interfaces-module>spring-boot-example-services-interfaces</services-interfaces-module>
         <persistence-module>spring-boot-example-persistence</persistence-module>
         <services-module>spring-boot-example-services</services-module>
         <webapi-module>spring-boot-example-webapi</webapi-module>
@@ -43,9 +43,9 @@
 
 
     <modules>
-        <module>${utils-module}</module>
-        <module>${models-module}</module>
-        <module>${interfaces-module}</module>
+        <module>${core-module}</module>
+        <module>${persistence-interfaces-module}</module>
+        <module>${services-interfaces-module}</module>
         <module>${persistence-module}</module>
         <module>${services-module}</module>
         <module>${webapi-module}</module>
