+++ spring-boot-example-core/src/main/java/com/bellotapps/examples/spring_boot_example/error_handling/helpers/ValidationHelper.java
