--- spring-boot-example-interfaces/pom.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-         xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.bellotapps.examples</groupId>
-        <artifactId>spring-boot-example</artifactId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>spring-boot-example-interfaces</artifactId>
-    <name>spring-boot-example-interfaces</name>
-
-    <properties>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
-
-    <dependencies>
-        <!-- Project modules -->
-        <dependency>
-            <groupId>com.bellotapps.examples</groupId>
-            <artifactId>spring-boot-example-models</artifactId>
-            <version>${project.parent.version}</version>
-        </dependency>
-
-        <!-- External Libraries -->
-        <!-- Spring data JPA for DAOs definitions -->
-        <dependency>
-            <groupId>org.springframework.data</groupId>
-            <artifactId>spring-data-jpa</artifactId>
-        </dependency>
-
-    </dependencies>
-</project>
