--- spring-boot-example-models/pom.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-         xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.bellotapps.examples</groupId>
-        <artifactId>spring-boot-example</artifactId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>spring-boot-example-models</artifactId>
-    <name>spring-boot-example-models</name>
-
-    <properties>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
-
-    <dependencies>
-        <!-- Project modules -->
-        <dependency>
-            <groupId>com.bellotapps.examples</groupId>
-            <artifactId>spring-boot-example-utils</artifactId>
-            <version>${project.parent.version}</version>
-        </dependency>
-
-        <!-- External Libraries -->
-        <!-- Hibernate Core in order to get JPA annotations (and some vendor specific features) -->
-        <dependency>
-            <groupId>org.hibernate</groupId>
-            <artifactId>hibernate-core</artifactId>
-        </dependency>
-
-        <!-- JUnit -->
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-        </dependency>
-    </dependencies>
-</project>
