+++ spring-boot-example-persistence-interfaces/src/main/java/com/bellotapps/examples/spring_boot_example/persistence/daos/UserDao.java
@@ -1,0 +1,46 @@
+package com.bellotapps.examples.spring_boot_example.persistence.daos;
+
+import com.bellotapps.examples.spring_boot_example.persistence.custom_repositories.ExtendedJpaRepository;
+import com.bellotapps.examples.spring_boot_example.models.User;
+
+import java.util.Optional;
+
+/**
+ * Defines behaviour of the DAO in charge of managing {@link User}s data.
+ */
+public interface UserDao extends ExtendedJpaRepository<User, Long> {
+
+    /**
+     * Retrieves the {@link User} with the given {@code username}.
+     *
+     * @param username The {@link User}'s username.
+     * @return A <b>nullable</b> {@link Optional} of {@link User}
+     * containing the {@link User} with the given {@code username} if it exists, or {@code null} otherwise.
+     */
+    Optional<User> findByUsername(String username);
+
+    /**
+     * Retrieves the {@link User} with the given {@code email}.
+     *
+     * @param email The {@link User}'s email.
+     * @return A <b>nullable</b> {@link Optional} of {@link User}
+     * containing the {@link User} with the given {@code email} if it exists, or {@code null} otherwise.
+     */
+    Optional<User> findByEmail(String email);
+
+    /**
+     * Checks if a {@link User} exists with the given {@code username}.
+     *
+     * @param username The username to check if a {@link User} exists with.
+     * @return {@code true} if a {@link User} exists with the given {@code username}, or {@code false} otherwise.
+     */
+    boolean existsByUsername(String username);
+
+    /**
+     * Checks if a {@link User} exists with the given {@code email}.
+     *
+     * @param email The email to check if a {@link User} exists with.
+     * @return {@code true} if a {@link User} exists with the given {@code email}, or {@code false} otherwise.
+     */
+    boolean existsByEmail(String email);
+}
