+++ spring-boot-example-webapi/src/main/java/com/bellotapps/examples/spring_boot_example/webapi/security/PasswordValidatorImpl.java
@@ -2,7 +2,7 @@ package com.bellotapps.examples.spring_boot_example.webapi.security;
 
 import com.bellotapps.examples.spring_boot_example.error_handling.errros.ValidationError;
 import com.bellotapps.examples.spring_boot_example.exceptions.ValidationException;
-import com.bellotapps.examples.spring_boot_example.interfaces.secutiry.PasswordValidator;
+import com.bellotapps.examples.spring_boot_example.security.PasswordValidator;
 import org.springframework.stereotype.Component;
 
 import java.util.Collections;
